+++ src/lexer/lexer.ts
@@ -1,24 +1,6 @@
-export class Token {
-    constructor(
-        public type: string,
-        public value: string,
-        public line: number,
-        public column: number
-    ) {}
-}
-
-export class LexerError extends Error {
-    constructor(message: string, public line: number, public column: number) {
-        super(message);
-        this.name = 'LexerError';
-    }
-}
-
-interface TokenDefinition {
-    type: string;
-    regex: RegExp;
-    ignore?: boolean;
-}
+import { Token } from './token';
+import { LexerError } from './lexerError';
+import { TokenDefinition, tokenDefinitions } from './tokenDefinitions';
 
 export class Lexer {
     private tokens: Token[] = [];
@@ -58,44 +40,20 @@ export class Lexer {
 
     private addToken(type: string, value: string): void {
         this.tokens.push(new Token(type, value, this.line, this.column));
-        this.column += value.length;
     }
+    
 
     private updatePosition(str: string): void {
         for (const char of str) {
             if (char === '\n') {
                 this.line++;
-                this.column = 1; 
+                this.column = 1;
             } else {
                 this.column++;
             }
         }
     }
 }
-
-const tokenDefinitions: TokenDefinition[] = [
-    { type: 'WHITESPACE', regex: /^\s+/, ignore: true },
-    { type: 'COMMENT', regex: /^\/\/.*/, ignore: true },
-    { type: 'MULTILINE_COMMENT', regex: /^\/\*[\s\S]*?\*\//, ignore: true },
-    { type: 'NUMBER', regex: /^\d+(\.\d+)?/ },
-    { type: 'STRING', regex: /^"([^"\\]|\\.)*"/ },
-    { type: 'IDENTIFIER', regex: /^[a-zA-Z_]\w*/ },
-    { type: 'EQUALS_EQUALS', regex: /^==/ },
-    { type: 'NOT_EQUALS', regex: /^!=/ },
-    { type: 'LESS_EQUALS', regex: /^<=/ },
-    { type: 'GREATER_EQUALS', regex: /^>=/ },
-    { type: 'EQUALS', regex: /^=/ },
-    { type: 'PLUS', regex: /^\+/ },
-    { type: 'MINUS', regex: /^-/ },
-    { type: 'ASTERISK', regex: /^\*/ },
-    { type: 'SLASH', regex: /^\// },
-    { type: 'LEFT_PAREN', regex: /^\(/ },
-    { type: 'RIGHT_PAREN', regex: /^\)/ },
-    { type: 'LEFT_BRACE', regex: /^\{/ },
-    { type: 'RIGHT_BRACE', regex: /^\}/ },
-    { type: 'SEMICOLON', regex: /^;/ },
-    { type: 'COMMA', regex: /^,/ },
-];
 
 export function lexer(input: string): Token[] {
     const lex = new Lexer(input, tokenDefinitions);
