+++ src/lexer/lexerError.ts
@@ -1,0 +1,6 @@
+export class LexerError extends Error {
+    constructor(message: string, public line: number, public column: number) {
+        super(message);
+        this.name = 'LexerError';
+    }
+}
