+++ src/lexer/token.ts
@@ -1,0 +1,8 @@
+export class Token {
+    constructor(
+        public type: string,
+        public value: string,
+        public line: number,
+        public column: number
+    ) {}
+}
